Update pt_BR.lang
I'll leave the "Tinker's" translation open - if someone wants to change it, fine by me.
"Portuguesation" of some ores, like Signalum and Enderium. I thought it would be better to do this because, if real ores have different names depending on the language, why not?
Again, open to changes. If someone thinks "Signalo" is better as "Sinalo" or something like that, change as you please.
Probably gonna change "Pyrotheum" and "Cryotheum". Have to think how it's gonna turn out though, because my present ideas are just too much wonky.

MrKunji mught want to change, for instance, "Ferrous" to "Nickel" or "Shiny" to "Platinum". I don't know if this is a good idea because it alters the real word. Don't know if TF have different names just 'coz or what, but, one more time, open to changes.

~I do fight for my translations, so that's why I talked so much about being "open to changes" :p